requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.3.9
appnope==0.1.0
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.1
camelot-py==0.8.0
cchardet==2.1.6
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
elasticsearch==7.7.1
et-xmlfile==1.0.1
fake-useragent==0.1.11
geographiclib==1.50
geopy==1.22.0
googletransx==2.4.2
idna==2.8
imutils==0.5.2
ipython==7.15.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.0
kiwisolver==1.2.0
matplotlib==3.2.1
multidict==4.7.6
numpy==1.18.4
opencv-python==4.1.1.26
openpyxl==3.0.3
pandas==1.0.4
parso==0.7.0
pdfminer.six==20200517
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.2
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycares==3.1.1
pycparser==2.20
pycryptodome==3.9.7
Pygments==2.6.1
pyparsing==2.4.7
PyPDF2==1.26.0
PySocks==1.7.1
pytesseract==0.2.6
pytextractor==0.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.22.0
schedule==0.6.0
six==1.15.0
sortedcontainers==2.1.0
soupsieve==2.0.1
traitlets==4.3.3
twint==2.1.20
urllib3==1.25.9
wcwidth==0.1.9
yarl==1.4.2
